Moved TestRow into its own file.
